// src/models/ChartOptions.model.ts
export interface ChartType {
    chart?: any;
    plotOptions?: any;
    colors?: any;
    series?: any;
    stroke?: any;
    fill?: any;
    labels?: any;
    markers?: any;
    legend?: any;
    xaxis?: any;
    yaxis?: any;
    tooltip?: any;
    grid?: any;
    type?: any;
    sparkline?: any;
    dataLabels?: any;
    height?: any;
    options?: any;
    toolbar?: any;
    stacked?: any;
    color?: any;
    width?: any;
    padding?: any;
    fixed?: any;
    marker?: any;
    responsive?: any;
    states?: any;
    title?: any;
    subtitle?: any;
  }